wasm: move main-template render into pkg/rt/wasm (fix root test ban)
The generated-main template, its -w-host-eval splice, and renderWasmMain
(now exported RenderMain) move from the root wasm.go into pkg/rt/wasm,
alongside AssembleHTML and the HTML/JS build assets they substitute into. Its
test moves with it, clearing the repo-root test-file ban that was failing CI
(scripts/check_test_location.py). Pure move — no behavior change; buildWasm now
calls wasmassets.RenderMain.
